Created new test file in GitApp/tests
